feat: Add user 'about' section and related icons
Introduces a new 'about' section to the UserProfile type, allowing for the storage of location, age, interests, website, and an extended bio. Also adds new SVG icons for 'LocationMarker', 'Cake', 'Sparkles', and 'Link' to support this new feature and enhances the Profile component to display and manage this information. The constants file is updated with initial 'about' data for a sample user.